+++ src/main/java/at/grueneis/spengergasse/lesson_plan/domain/BasePersistable.java
@@ -19,15 +19,9 @@ public abstract class BasePersistable {
 	
 
 	private Long id;
-    private String md5Hash;
-    
-    public String getMd5Hash() {
-		return md5Hash;
-	}
 
-    public void setMd5Hash(String md5Hash) {
-		this.md5Hash = md5Hash;
-	}
+    
+
     
     public Long getId() {
         return id;
@@ -64,11 +58,5 @@ public abstract class BasePersistable {
     public String toString() {
         return getClass().getSimpleName() + "[id=" + id + "]";
     }
-    
-    /**
-     * -) If the attribute is an object reference, the value in the string should be the ID of the referenced object.
-     * -) If the attribute is a list containing object references , the value should be the IDs of all elements in the list
-     * @return String[]: Returns a Stringarray containing the values of all attributes. (1 element per attribute). 
-     */
-    public abstract String[] getAllAttributesAsString();
+
 }
